rename: feature auth 폴더 정리
