+++ src/features/auth/index.ts
@@ -1,5 +1,5 @@
 export * from './ui';
-export * from './form.hook';
-export * from './form.types';
-export * from './form.utils';
-export * from './progress.type';
+export * from './hook/form.hook';
+export * from './model/form.types';
+export * from './lib/form.utils';
+export * from './model/progress.type';
